+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changes before Tatum release are not documented in this file.
 
 #### Added
 
+- Transaction timeouts in Autostaker (https://github.com/streamr-dev/network/pull/3236)
+
 #### Changed
 
 #### Deprecated